+++ environment.js
@@ -4,8 +4,8 @@ export const environmentSettings = { // Here begins the object
   //This is a list of variable properties for skills
   //Always use all lowercase for everything!
 
-  "jsk": true, //Show skills and rules from the jumpstart kit
-  "core": false //Show skills and rules from the core rulebook for actual play
+  "jsk": false, //Show skills and rules from the jumpstart kit
+  "core": true //Show skills and rules from the core rulebook for actual play
 
   //Don't change anything below this line